Create killvm.sh
When executed on the Proxmox server via the CLI, this script can simplify force-stopping a VM when it is frozen rather than restarting the entire server.

Typically, the GUI and CLI commands are prevented from functioning because of a lock on the proc file. This script tests for that lock, then brings down the VM. If there is no lock on the file, the user is given the option to execute the kill command anyway.